modifica controller
